[EAM-2286] Reset error messages state
Since the dialog is being rendered by the `Activities` and `PartUsage`
(parent) components, the custom hook state gets associated to them.
This means that only when the Activities or PartUsage panel get
unmounted, does the custom hook's state gets cleared as well.
Thus, we explicitly have to reset the errorMessages state so that we do
not get the same old error messages upon re-opening the dialogs.