[EAM-2286] Update way of resetting values
Changed from using the cleanup function on useEffect to just checking
that the dialog is meant to be closed. This is more in-line with what is
expected since the unmount was being used as if the Dialog component
itselft was being unmounted where in fact we just want to check when
it is closed.