[EAM-2286] Reset error messages state (#176)
* [EAM-2286] Reset error messages state

Since the dialog is being rendered by the `Activities` and `PartUsage`
(parent) components, the custom hook state gets associated to them.
This means that only when the Activities or PartUsage panel get
unmounted, does the custom hook's state gets cleared as well.
Thus, we explicitly have to reset the errorMessages state so that we do
not get the same old error messages upon re-opening the dialogs.

* [EAM-2286] Fix mui-related warnings

* [EAM-2286] Update way of resetting values

Changed from using the cleanup function on useEffect to just checking
that the dialog is meant to be closed. This is more in-line with what is
expected since the unmount was being used as if the Dialog component
itselft was being unmounted where in fact we just want to check when
it is closed.